+++ vpc-20160428/alibabacloud_vpc20160428/client.py
@@ -3477,6 +3477,8 @@ class Client(OpenApiClient):
             query['ResourceOwnerId'] = request.resource_owner_id
         if not UtilClient.is_unset(request.secondary_cidr_block):
             query['SecondaryCidrBlock'] = request.secondary_cidr_block
+        if not UtilClient.is_unset(request.secondary_cidr_mask):
+            query['SecondaryCidrMask'] = request.secondary_cidr_mask
         if not UtilClient.is_unset(request.vpc_id):
             query['VpcId'] = request.vpc_id
         req = open_api_models.OpenApiRequest(
@@ -3538,6 +3540,8 @@ class Client(OpenApiClient):
             query['ResourceOwnerId'] = request.resource_owner_id
         if not UtilClient.is_unset(request.secondary_cidr_block):
             query['SecondaryCidrBlock'] = request.secondary_cidr_block
+        if not UtilClient.is_unset(request.secondary_cidr_mask):
+            query['SecondaryCidrMask'] = request.secondary_cidr_mask
         if not UtilClient.is_unset(request.vpc_id):
             query['VpcId'] = request.vpc_id
         req = open_api_models.OpenApiRequest(
@@ -12686,6 +12690,8 @@ class Client(OpenApiClient):
             query['DryRun'] = request.dry_run
         if not UtilClient.is_unset(request.enable_ipv_6):
             query['EnableIpv6'] = request.enable_ipv_6
+        if not UtilClient.is_unset(request.ipv_4cidr_mask):
+            query['Ipv4CidrMask'] = request.ipv_4cidr_mask
         if not UtilClient.is_unset(request.ipv_4ipam_pool_id):
             query['Ipv4IpamPoolId'] = request.ipv_4ipam_pool_id
         if not UtilClient.is_unset(request.ipv_6cidr_block):
@@ -12762,6 +12768,8 @@ class Client(OpenApiClient):
             query['DryRun'] = request.dry_run
         if not UtilClient.is_unset(request.enable_ipv_6):
             query['EnableIpv6'] = request.enable_ipv_6
+        if not UtilClient.is_unset(request.ipv_4cidr_mask):
+            query['Ipv4CidrMask'] = request.ipv_4cidr_mask
         if not UtilClient.is_unset(request.ipv_4ipam_pool_id):
             query['Ipv4IpamPoolId'] = request.ipv_4ipam_pool_id
         if not UtilClient.is_unset(request.ipv_6cidr_block):